// FoodDiary.Web.Client/projects/fooddiary-admin/src/app/shared/api/sdk/generated/model/admin-retention-cohort-http-response.ts
/**
 * FoodDiary Admin API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface AdminRetentionCohortHttpResponse {
    activatedWithinSevenDays?: number | null;
    date?: string;
    day1?: number | null;
    day30?: number | null;
    day7?: number | null;
    registered?: number;
}