// FoodDiary.Web.Client/projects/fooddiary-admin/src/app/shared/api/sdk/generated/model/admin-retention-report-http-response.ts
/**
 * FoodDiary Admin API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { AdminRetentionCohortHttpResponse } from './admin-retention-cohort-http-response';
import { AdminRetentionDayHttpResponse } from './admin-retention-day-http-response';

export interface AdminRetentionReportHttpResponse {
    activeUsersInPeriod?: number;
    activityByDay?: Array<AdminRetentionDayHttpResponse>;
    asOfUtc?: string;
    cohortFromUtc?: string | null;
    cohortToUtc?: string | null;
    cohorts?: Array<AdminRetentionCohortHttpResponse>;
    fromUtc?: string;
    mealEntriesInPeriod?: number;
    toUtc?: string;
}